Initialization(3.0)

+++ Lynda1-Exploring lake pend oreille/src/Macros.java
@@ -1,9 +1,9 @@
 public class Macros {
-	public static final double AIR_TEMP = 21;
-	public static final double BAROMETRIC_PRESSURE = 29;
-	public static final double DEW_POINT = 37;
-	public static final double RELATIVE_HUMIDITY = 45;
-	public static final double WIND_DIRECTION = 53;
-	public static final double WIND_GUST = 61;
-	public static final double WIND_SPEED = 69;
+	public static final Integer AIR_TEMP = 21;
+	public static final Integer BAROMETRIC_PRESSURE = 29;
+	public static final Integer DEW_POINT = 37;
+	public static final Integer RELATIVE_HUMIDITY = 45;
+	public static final Integer WIND_DIRECTION = 53;
+	public static final Integer WIND_GUST = 61;
+	public static final Integer WIND_SPEED = 69;
 }
